requirements.txt
black
cookiecutter
isort
pytest
pytest-cookies
pytest-jsonschema >= 1.0.0a2
cookieplone>=0.4.1
GitPython
wheel
